Shift AdvancedJMM samples
