+++ jcstress-samples/src/main/java/org/openjdk/jcstress/samples/jmm/advanced/AdvancedJMM_07_ArrayVolatility.java
@@ -35,11 +35,11 @@ import java.lang.invoke.VarHandle;
 
 import static org.openjdk.jcstress.annotations.Expect.*;
 
-public class AdvancedJMM_06_ArrayVolatility {
+public class AdvancedJMM_07_ArrayVolatility {
 
     /*
         How to run this test:
-            $ java -jar jcstress-samples/target/jcstress.jar -t AdvancedJMM_06_ArrayVolatility[.SubTestName]
+            $ java -jar jcstress-samples/target/jcstress.jar -t AdvancedJMM_07_ArrayVolatility[.SubTestName]
      */
 
     /*
