+++ jcstress-samples/src/main/java/org/openjdk/jcstress/samples/jmm/advanced/AdvancedJMM_10_WrongAcquireOrder.java
@@ -34,32 +34,27 @@ import static org.openjdk.jcstress.annotations.Expect.*;
 
 @JCStressTest
 @State
-@Outcome(id = {"0, 0", "1, 1"}, expect = ACCEPTABLE,             desc = "Boring")
-@Outcome(id = "0, 1",           expect = ACCEPTABLE,             desc = "Plausible")
-@Outcome(id = "1, 0",           expect = ACCEPTABLE_INTERESTING, desc = "Whoa")
-public class AdvancedJMM_10_WrongAcquireReleaseOrder {
+@Outcome(id = {"0, 0", "1, 1"}, expect = ACCEPTABLE, desc = "Boring")
+@Outcome(id = {"0, 1", "1, 0"}, expect = ACCEPTABLE, desc = "Plausible")
+public class AdvancedJMM_10_WrongAcquireOrder {
 
     /*
         How to run this test:
-            $ java -jar jcstress-samples/target/jcstress.jar -t AdvancedJMM_10_WrongAcquireReleaseOrder
+            $ java -jar jcstress-samples/target/jcstress.jar -t AdvancedJMM_10_WrongAcquireOrder
      */
 
     /*
-       ----------------------------------------------------------------------------------------------------------
+      ----------------------------------------------------------------------------------------------------------
 
-        This example effectively inverts the BasicJMM_06_Causality test: it "guards" the volatile "g" with
-        plain "x". The interesting outcome (1, 0) cannot be explained by the sequential execution of this code,
-        nevertheless it is allowed by JMM, because it is a data race.
-
-        Note that even if we have seen "x = 1", we can still see "g = 1". That is, adding "volatile" to "g"
-        does not guarantee the _promptness_ of publishing of "g".
+        For completeness, the example that has a wrong acquire order. All these results can be explained by
+        sequential execution of the code.
 
         x86_64:
           RESULT        SAMPLES     FREQ       EXPECT  DESCRIPTION
-            0, 0  2,694,178,631   58.36%   Acceptable  Boring
-            0, 1     69,859,185    1.51%   Acceptable  Plausible
-            1, 0        930,435    0.02%  Interesting  Whoa
-            1, 1  1,851,647,173   40.11%   Acceptable  Boring
+            0, 0  2,560,656,086   55.33%   Acceptable  Boring
+            0, 1      2,961,349    0.06%   Acceptable  Plausible
+            1, 0      7,885,064    0.17%   Acceptable  Plausible
+            1, 1  2,056,684,125   44.44%   Acceptable  Boring
      */
 
     int x;
@@ -67,13 +62,13 @@ public class AdvancedJMM_10_WrongAcquireReleaseOrder {
 
     @Actor
     public void actor1() {
-        g = 1;
         x = 1;
+        g = 1;
     }
 
     @Actor
     public void actor2(II_Result r) {
         r.r1 = x;
-        r.r2 = g;
+        r.r2 = g; // acquiring too late
     }
 }
